chore: rename Shared to shared
