+++ src/app/styles/base/settings.css
@@ -74,6 +74,6 @@ img {
 
 .disabled-link {
   pointer-events: none;
-  opacity: 0.5;
   cursor: default;
+  opacity: 0.5;
 }
