+++ src/shared/ui/Icon/CartIcon.tsx
