+++ src/shared/ui/Icon/ProfileIcon.tsx
