+++ src/shared/ui/Icon/WishlistIcon.tsx
