+++ src/shared/ui/Icon/index.ts
