+++ src/shared/ui/Logo/index.ts
