+++ src/shared/validation/emailValidation.ts
